requirements.txt
configargparse==1.5.3
rich==10.15.2
setuptools~=59.5.0
PyYAML~=6.0
schema~=0.7.5
psycopg2==2.9.2
SQLAlchemy~=1.4.27
faker==10.0.0
